# swagger-first.yaml
---
swagger: "2.0"
info:
  description: "Descricao da API de pagamento"
  version: "1.0.0"
  title: "Pagamento API"
host: "localhost:8080"
basePath: "/nataly-api/v1"
schemes:
- "http"
paths:
  /pagamentos:
    post:
      summary: "Cria novo pagamento"
      description: "Endpoint de pagamento"
      operationId: "criarPagamento"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "transacao"
        description: "Transação"
        required: true
        schema:
          $ref: "#/definitions/Transacao"
      responses:
        201:
          description: "Novo pagamento criado"
          schema:
            $ref: "#/definitions/Pagamento"
          headers:
            Location:
              type: "string"
              description: "uri do novo pagamento"
  /pagamentos/{id}:
    get:
      operationId: "buscaPagamento"
      description: "|
        O Pagamentos delete endpoint *busca o* pagamento."
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Pagamento"
          headers: {}
    put:
      summary: "Confirma o pagamento"
      operationId: "confirmarPagamento"
      description: "|
        O Pagamentos put endpoint *confirma o* pagamento."
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Pagamento confirmado"
          schema:
            $ref: "#/definitions/Pagamento"
    delete:
      summary: Cancela um pagamento
      operationId: "cancelarPagamento"
      description: "|
        O Pagamentos delete endpoint *cancela o* pagamento."
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Pagamento cancelado"
          schema:
            $ref: "#/definitions/Pagamento"
definitions:
  Transacao:
    type: "object"
    properties:
      codigo:
        type: "string"
        description: Codigo da transacao.
      titular:
        type: "string"
        description: Titular da transacao.
      data:
        type: "string"
        format: "date"
        description: Data da transacao.
      valor:
        type: "number"
        format: "double"
        description: Valor da transacao.
  Pagamento:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
        description: "Id do pagamento."
      status:
        type: "string"
        description: "Status do pagamento"
      valor:
        type: "number"
        format: "double"
        description: "Valor do pagamento."